renames TelemetryAppBar.svelte to GLAMMark.svelte
